+++ Commands/VideoCommandMethods.cs
@@ -20,12 +20,12 @@ namespace ooceBot.Commands
                 {
                     OBSWebsocket websocket = await OBSManager.ConnectToOBSWebsocket();
 
-                    // Get the current volume from the API
-                    double originalVolume = await VolumeControl.GetNightbotCurrentVolume(args.NightbotSongRequestClient);
-                    double volumeChange = originalVolume * 0.9;
+                    // Get the current volume from the API (whole numbers are the only accepted values, so we use ints for all calculations)
+                    int originalVolume = await VolumeControl.GetNightbotCurrentVolume(args.NightbotSongRequestClient);
+                    int volumeChange = (int)(originalVolume * 0.9);
 
                     // Keep track of the volume for the reset after the video is done
-                    double updatedVolume = await VolumeControl.ReduceVolume(args.NightbotSongRequestClient, originalVolume, volumeChange);
+                    int updatedVolume = await VolumeControl.ReduceVolume(args.NightbotSongRequestClient, originalVolume, volumeChange);
 
                     // The scene needs to exist in the currently selected scene, so fetch the current scene name and its items
                     var currentScene = websocket.GetCurrentProgramScene();
@@ -54,12 +54,12 @@ namespace ooceBot.Commands
                 {
                     OBSWebsocket websocket = await OBSManager.ConnectToOBSWebsocket();
 
-                    // Get the current volume from the API
-                    double originalVolume = await VolumeControl.GetNightbotCurrentVolume(args.NightbotSongRequestClient);
-                    double volumeChange = originalVolume * 0.9;
+                    // Get the current volume from the API (whole numbers are the only accepted values, so we use ints for all calculations)
+                    int originalVolume = await VolumeControl.GetNightbotCurrentVolume(args.NightbotSongRequestClient);
+                    int volumeChange = (int)(originalVolume * 0.9);
 
                     // Keep track of the volume for the reset after the video is done
-                    double updatedVolume = await VolumeControl.ReduceVolume(args.NightbotSongRequestClient, originalVolume, volumeChange);
+                    int updatedVolume = await VolumeControl.ReduceVolume(args.NightbotSongRequestClient, originalVolume, volumeChange);
 
                     // The scene needs to exist in the currently selected scene, so fetch the current scene name and its items
                     var currentScene = websocket.GetCurrentProgramScene();
@@ -87,12 +87,12 @@ namespace ooceBot.Commands
                 {
                     OBSWebsocket websocket = await OBSManager.ConnectToOBSWebsocket();
 
-                    // Get the current volume from the API
-                    double originalVolume = await VolumeControl.GetNightbotCurrentVolume(args.NightbotSongRequestClient);
-                    double volumeChange = originalVolume * 0.9;
+                    // Get the current volume from the API (whole numbers are the only accepted values, so we use ints for all calculations)
+                    int originalVolume = await VolumeControl.GetNightbotCurrentVolume(args.NightbotSongRequestClient);
+                    int volumeChange = (int)(originalVolume * 0.9);
 
                     // Keep track of the volume for the reset after the video is done
-                    double updatedVolume = await VolumeControl.ReduceVolume(args.NightbotSongRequestClient, originalVolume, volumeChange);
+                    int updatedVolume = await VolumeControl.ReduceVolume(args.NightbotSongRequestClient, originalVolume, volumeChange);
 
                     // The scene needs to exist in the currently selected scene, so fetch the current scene name and its items
                     var currentScene = websocket.GetCurrentProgramScene();
